Enable a working mvp.json file to generate
Some of the routes files had syntax errors in them which had to be
corrected to avoid errors on generation.

The questionnaire file did not include some new sections so required
to be updated in order to generate.

Several new sections included in the routes were added to avoid
errors.

Gherkins folder was renamed to features. This was done to comply
with the cucumber standard.