Convert sections to JavaScript files
This commit has changed each json file to a javascipt file.
This was done to allow the data to be better maintained thru the use
of syntax such as multiline strings.

In order to facilitate this, each file's JSON object is now exported
using module.exports. Each JSON can now be fetched using `require`.