Fix confirmation page (#7)
* Fix confirmation page

This commit fixes the confirmation page. The appropriate html classes
have now been added to make the application mirror the prototype.

This commit also establishes a method of piping data into a section
This is done by entering a json pointer enclosed within double pipe
symbols. The pointer refers to a value within the questionnaire
and the DCS will recognise this syntax and substitute the value.